refactor: split trait flattening
